+++ CHANGES.md
@@ -22,6 +22,7 @@ post, put, etc.
 - Added detection for timeout when reading data from client.
 - Added property `host`, `path`, `query`, `queryString`, `https`, `ip`,
 `aborted`, `closed`, etc for ServerRequest object.
+- Added HTTP Cookies encoder.
 
 ## v0.1.0-b1